Rename Test to MCCTeleOp, make it an OpMode
In this case I think it makes more sense for the TeleOp class to be an OpMode
instead of LinearOpMode because the loop method is convenient and clearer than
doing a while opModeIsActive().

Also, I added the basic structure of the TeleOp class, where there is a method
for each subassembly which is used to do different actions depending on input
from the gamepads.

There is also an updateTelemetry() call at the end of loop() which will be used
for tuning/testing.