Add mecanum drive code to teleop
Add mecanum drive code in the drive() method according to GM0 mecanum tutorial
and Ether paper. This includes the basic calculation to convert gamepad inputs
into power for each motor, squaring or cubing the inputs (unless the drive
mode is normal), and scaling the normal/squared/cubed inputs to be within
[-1, 1]. I also followed the advice of GM0's tutorial and multiplied the
horizontal (strafe) gamepad input by a constant which we can tune later
according to the drivers' preferences. This is supposed to counteract
strafing imperfections.

Create DriveMode enum to allow easy switching between normal, squared,
and cubed drive. This is a separate file because we will probably end up
having multiple teleop classes, so it's better to change it in one place
and import instead of copying and pasting the enum for each teleop class we make.

Add todo item to Drivetrain.java to remember to reverse motor directions if needed