Make gamepad 2 B button stop the flywheel
Having gamepad 2 A button assigned to both starting and stopping
the flywheel caused inconsistent results. Sometimes pressing it
once wasn't enough to start it. Sometimes pressing it once wasn't
enough to stop it. At the end of the day, I decided it's much better
in a match to have a consistent, separate button to stop the flywheel.